feat: create and implement intial NavRoot()
